Add parameter 'BINARY_DIR'.
Background:
ssh-ident needs to determine which binary to run next
to complete the requested command. Generally, this binary
is 'ssh'.

With previous changes (thanks to black2754), code was added
to detect the binary to run based on argv[0], and to find
it in PATH.

Depending on the setup, this can be fragile as argv[0] may
not contain a path at all, or the path in $PATH may contain
symlinks or bind mounts, which would be difficult to exclude.

In this change:
- if BINARY_DIR is set, look for the next binary to run
  in that directory only. This could be set to "/usr/bin"
  on most unix systems.
  This is optional, the default is still to search in
  the path.

Additionally:
- be a bit more paranoid when searching in PATH. Specifically,
  handle relative paths without defaulting to the CWD.
- add more verbose messages, as those problems may be difficult
  or obscure to debug.