Fix compatibility issues with SSH_ASKPASS.
Background:

Previously, ssh-agent was run in an empty environment.
This caused the agent to fail when a key was added
with ssh-add -c and SSH_ASKPASS is set.

When ssh-add was called with SSH_ASKPASS, the
passphrase was still read from the terminal and the
askpass program was not called.

In this commit

- ssh-ident detects if SSH_ASKPASS is set and invokes
  ssh-add with </dev/null to force askpass to be
  used instead of terminal input.
- ssh-agent gets started with specific environment
  variables to be able to call SSH_ASKPASS.